requirements.txt
numpy
pandas
alpaca-trade-api
python-dotenv